quick modalをfeature化
