+++ apps/client/src/features/modal/index.ts
@@ -1,0 +1,4 @@
+/** @public */
+export { confirm } from "./model";
+/** @public */
+export { QuickModalContainer } from "./ui";
