+++ apps/client/src/features/modal/model/fns.ts
@@ -1,0 +1,4 @@
+import { openModal } from "./store";
+
+export const confirm = (message: string): Promise<boolean> =>
+  openModal("confirm", { message });
