+++ apps/client/src/features/modal/model/index.ts
@@ -1,0 +1,2 @@
+export { confirm } from "./fns";
+export { type ModalInput, type Modals, modal } from "./store";
