+++ apps/client/src/features/modal/ui/index.ts
@@ -1,0 +1,1 @@
+export { QuickModalContainer } from "./container";
