--- apps/client/src/modal/container.ts
@@ -1,1 +1,0 @@
-export * from "./quick/container";
