--- apps/client/src/modal/index.ts
@@ -1,1 +1,0 @@
-export * from "./quick";
