--- apps/client/src/modal/quick/index.ts
@@ -1,6 +1,0 @@
-import { openQuickModal } from "./store";
-
-export * from "./container";
-
-export const confirm = (message: string): Promise<boolean> =>
-    openQuickModal("confirm", { message });
